Add a custom parser to handle Sendgrid's 500 responses
When Sendgrid returns a 502 they return it with a HTML error
page for the body, even though we requested application/json.
HTTParty will try to parse this response as JSON and then
throw a parsing error, which hides the real response that we
recieved from Sendgrid. This fix adds a custom parser which
catches the JSON:ParseError and falls back to returning
the body as it was given along with an error.

This is the recommended solution by HTTParty devs:
https://github.com/jnunemaker/httparty/issues/78#issuecomment-920787